add If's

+++ src/controllers/deleteBeer.ts
@@ -5,14 +5,20 @@ import { PrismaClient } from "@prisma/client";
 const prisma = new PrismaClient();
 
 export const deleteBeer = async (req: Request, res: Response) => {
+
     try {
+        const id = req.params.id
+        if (!id) {
+            return res.status(400).json({ message: 'ID is required' });
+        }
         await prisma.beer.delete({
             where: {
-                id: req.params.id
-            }     
+                id: id
+            }
         });
         res.json({ message: 'Beer deleted' })
     } catch (error) {
-        res.status(500).json({ message: 'ID not found' })
+        console.error(error)
+        res.status(500).json({ message: 'ID not found' });
     }
 };
